CleanUp of project names and namespaces
