--- src/MLib3.Protocols.Measurements.Abstractions/IProtocolDeserializer.cs
@@ -1,8 +1,0 @@
-using FluentResults;
-
-namespace MLib3.Protocols.Measurements;
-
-public interface IProtocolDeserializer
-{
-    Result<IProtocol> Deserialize(string data);
-}
