--- src/MLib3.Protocols.Measurements.Abstractions/IProtocolFileReader.cs
@@ -1,8 +1,0 @@
-using FluentResults;
-
-namespace MLib3.Protocols.Measurements;
-
-public interface IProtocolFileReader
-{
-    Result<IProtocol> Read(string filename);
-}
