--- src/MLib3.Protocols.Measurements.Abstractions/IProtocolFileWriter.cs
@@ -1,8 +1,0 @@
-﻿using FluentResults;
-
-namespace MLib3.Protocols.Measurements;
-
-public interface IProtocolFileWriter
-{
-    Result Write(string filename, IProtocol protocol);
-}
