--- src/MLib3.Protocols.Measurements.Abstractions/IProtocolSerializer.cs
@@ -1,8 +1,0 @@
-using FluentResults;
-
-namespace MLib3.Protocols.Measurements;
-
-public interface IProtocolSerializer
-{
-    Result<string> Serialize(IProtocol protocol);
-}
