--- src/MLib3.Protocols.Measurements.Json/MLib3.Protocols.Measurements.Json.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\MLib3.Protocols.Measurements.Abstractions\MLib3.Protocols.Measurements.Abstractions.csproj" />
-      <ProjectReference Include="..\MLib3.Protocols.Measurements.Serialization\MLib3.Protocols.Measurements.Serialization.csproj" />
-      <ProjectReference Include="..\MLib3.Protocols.Measurements\MLib3.Protocols.Measurements.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="FluentResults" Version="3.16.0" />
-    </ItemGroup>
-
-</Project>
