+++ src/MLib3.Protocols.Measurements.Serialization.Xml/Converters/IXmlVersionConverter.cs
@@ -1,0 +1,6 @@
+namespace MLib3.Protocols.Measurements.Serialization.Xml.Converters;
+
+internal interface IXmlVersionConverter
+{
+    string Convert(string xml);
+}
