+++ src/MLib3.Protocols.Measurements.Serialization/IProtocolDeserializer.cs
@@ -1,0 +1,8 @@
+using FluentResults;
+
+namespace MLib3.Protocols.Measurements.Serialization;
+
+public interface IProtocolDeserializer
+{
+    Result<IProtocol> Deserialize(string data);
+}
