+++ src/MLib3.Protocols.Measurements.Serialization/IProtocolSerializer.cs
@@ -1,0 +1,8 @@
+using FluentResults;
+
+namespace MLib3.Protocols.Measurements.Serialization;
+
+public interface IProtocolSerializer
+{
+    Result<string> Serialize(IProtocol protocol);
+}
