--- src/MLib3.Protocols.Measurements.Xml/Converters/IXmlVersionConverter.cs
@@ -1,6 +1,0 @@
-namespace MLib3.Protocols.Measurements.Xml.Converters;
-
-internal interface IXmlVersionConverter
-{
-    string Convert(string xml);
-}
