+++ tests/MLib3.Protocols.Measurements.Serialization.Json.UnitTests/MLib3.Protocols.Measurements.Serialization.Json.UnitTests.csproj
@@ -1,0 +1,30 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+
+        <IsPackable>false</IsPackable>
+        <IsTestProject>true</IsTestProject>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="coverlet.collector" Version="6.0.0"/>
+        <PackageReference Include="FluentAssertions" Version="7.0.0-alpha.4" />
+        <PackageReference Include="FluentResults.Extensions.FluentAssertions" Version="2.1.2" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0"/>
+        <PackageReference Include="xunit" Version="2.5.3"/>
+        <PackageReference Include="xunit.runner.visualstudio" Version="2.5.3"/>
+    </ItemGroup>
+
+    <ItemGroup>
+        <Using Include="Xunit"/>
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\src\MLib3.Protocols.Measurements.Serialization.Json\MLib3.Protocols.Measurements.Serialization.Json.csproj" />
+      <ProjectReference Include="..\MLib3.Protocols.Measurements.TestCore\MLib3.Protocols.Measurements.TestCore.csproj" />
+    </ItemGroup>
+
+</Project>
