+++ tests/MLib3.Protocols.Measurements.Serialization.Xml.UnitTests/Common/XmlSchemaValidatorV2.cs
@@ -1,0 +1,18 @@
+using System.Xml;
+using System.Xml.Linq;
+using System.Xml.Schema;
+
+namespace MLib3.Protocols.Measurements.Serialization.Xml.UnitTests.Common;
+
+public static class XmlSchemaValidatorV2
+{
+    public static Result Validate(string xml)
+    {
+        var schemas = new XmlSchemaSet();
+        schemas.Add("", XmlReader.Create(new StringReader(File.ReadAllText("./SchemaV2.xsd"))));
+        
+        var validationErrors = new List<string>();
+        XDocument.Parse(xml).Validate(schemas, (sender, args) => validationErrors.Add(args.Message));
+        return validationErrors.Any() ? Result.Fail(validationErrors) : Result.Ok();
+    }
+}
