+++ tests/MLib3.Protocols.Measurements.Serialization.Xml.UnitTests/SchemaV2.xsd
