+++ tests/MLib3.Protocols.Measurements.Serialization.Xml.UnitTests/SchemaV3.xsd
