--- tests/MLib3.Protocols.Measurements.Xml.UnitTests/MLib3.Protocols.Measurements.Xml.UnitTests.csproj
@@ -1,42 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-
-        <IsPackable>false</IsPackable>
-        <IsTestProject>true</IsTestProject>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="coverlet.collector" Version="6.0.0"/>
-        <PackageReference Include="FluentAssertions" Version="6.12.1" />
-        <PackageReference Include="FluentResults" Version="3.16.0" />
-        <PackageReference Include="FluentResults.Extensions.FluentAssertions" Version="2.1.2" />
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.1" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0"/>
-        <PackageReference Include="xunit" Version="2.5.3"/>
-        <PackageReference Include="xunit.runner.visualstudio" Version="2.5.3"/>
-    </ItemGroup>
-
-    <ItemGroup>
-        <Using Include="Xunit"/>
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\src\MLib3.Protocols.Measurements.Xml\MLib3.Protocols.Measurements.Xml.csproj" />
-      <ProjectReference Include="..\..\src\MLib3.Protocols.Measurements\MLib3.Protocols.Measurements.csproj" />
-      <ProjectReference Include="..\MLib3.Protocols.Measurements.TestCore\MLib3.Protocols.Measurements.TestCore.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Update="SchemaV2.xsd">
-        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      </None>
-      <None Update="SchemaV3.xsd">
-        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      </None>
-    </ItemGroup>
-
-</Project>
